增加导入导出资源重定位节的解析和测试对象

+++ per/COFFheader.h
@@ -59,7 +59,7 @@ public:
 	  {
 
 	  }
-	  BOOL Init(PCHAR base,long size,PETYPE petype)
+	  BOOL Init(PCHAR base,DWORD size,PETYPE petype)
 	  {
 		  m_petype=petype;
 		  assert(size>=20,L"文件长度不够(>=20)无法初始化COFFHeader",FALSE);
