+++ per/MDOSStub.h
@@ -9,7 +9,7 @@ public:
 
 	}
 
-	BOOL Init(PCHAR base,long size)
+	BOOL Init(PCHAR base,DWORD size)
 	{
 		assert(size>=0x3c+2,L"文件长度不够(>=0x3c+2),无法初始化MSDOSStub",FALSE);
 		m_peheader_off=*(PDWORD(base+0x3c));
